Add support for extra flags which will be passed to g++
EXTRA_FLAGS have been added. These are flags that don't control the g+++ script and will be passed untouched to g++ (for example, -d). They are stored in the configuration file and loaded from it. The compilation command now includes these flags, and their values ​​are displayed in the output.